v1.8.8: FloatingWindow pause state display + context menu fix for preset tag deletion inconsistency

+++ ME/Views/TimeTrackView.xaml.cs
@@ -384,7 +384,7 @@ namespace ME.Views
                 editItem.Click += (s, ev) => EditTag(tag);
                 menu.Items.Add(editItem);
 
-                if (!tag.IsDefault)
+                if (!tag.IsDefault && !tag.IsPreset)
                 {
                     var deleteItem = new MenuItem
                     {
